Refactor Loader.stories.ts to Loader.stories.tsx and add WithCustomIcon story
